Update and Add README
Note that I did this project - according to my pc - in Jan and I had
written this README as help.txt back then as can also be seen in the
previous commit(I had written it in print commands! must have been bored
AF).  So if it feels a bit off the ways of github, that's why.